requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
fer==22.5.1
pillow==10.1.0
numpy==1.26.2
moviepy==1.0.3

# Deep Learning - Fine-tuning
torch>=2.0.0
torchvision>=0.15.0
tqdm>=4.65.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0

# Emotion Detection Models
hsemotion>=0.3
deepface>=0.0.79
tf-keras>=2.15.0
